adapt all logs flow to mysql

+++ microservicios/db/src/migrations/1588901055731-CreateTableLog.ts
@@ -9,7 +9,7 @@ export class CreateTableLog1588901055731 implements MigrationInterface {
                 operation varchar(255) NOT NULL,\
                 operationType varchar(255) NOT NULL,\
                 payload varchar(255) NULL,\
-                unixStartTime int NOT NULL,\
+                unixStartTime BIGINT NOT NULL,\
                 executionTime int NOT NULL,\
                 resultPayload varchar(255) NULL,\
                 PRIMARY KEY (id)\
